requirements.txt
# Bot Framework SDK
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Azure Authentication
msal==1.24.0
azure-identity==1.14.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# JSON Web Tokens
PyJWT==2.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Linting & Formatting
black==23.11.0
flake8==6.1.0
